* 'develop' of https://github.com/CenterForOpenScience/osf.io: (105 commits)
  Fix test broken by merge
  Fix logic error in archive_tree_finished()
  Remove unused import
  add explicit dep on oauthlib
  Add contraints file arg to wheelhouse command
  Prevent spammy registrations from getting approved
  Make user list read-only and disable write tests
  Ensure analytics task for institutions is actually run each night
  jshint
  recaptcha optimization
  reset captcha on submission error
  fix test
  recaptcha validation and tests
  legacy form push state for inavlid captcha
  merge settings
  Backend implementation for recaptcha.
  recaptcha frontend, check public nodes setting
  Update version and changelog for version
  admin display users as links in list
  ban additional domains
  ...

# Conflicts:
#	website/routes.py

+++ website/search/elastic_search.py
@@ -349,7 +349,7 @@ def update_node(node, index=None, bulk=False, async=False):
     for file_ in paginated(OsfStorageFile, Q('node', 'eq', node)):
         update_file(file_, index=index)
 
-    if node.is_deleted or not node.is_public or node.archiving or node.is_spammy:
+    if node.is_deleted or not node.is_public or node.archiving or (node.is_spammy and settings.SPAM_FLAGGED_REMOVE_FROM_SEARCH):
         delete_doc(node._id, node, index=index)
     else:
         category = get_doctype_from_node(node)
@@ -447,7 +447,6 @@ def update_user(user, index=None):
 
 @requires_search
 def update_file(file_, index=None, delete=False):
-
     index = index or INDEX
 
     if not file_.node.is_public or delete or file_.node.is_deleted or file_.node.archiving:
@@ -469,9 +468,14 @@ def update_file(file_, index=None, delete=False):
     )
     node_url = '/{node_id}/'.format(node_id=file_.node._id)
 
+    guid_url = None
+    file_guid = file_.get_guid(create=False)
+    if file_guid:
+        guid_url = '/{file_guid}/'.format(file_guid=file_guid._id)
     file_doc = {
         'id': file_._id,
         'deep_url': file_deep_url,
+        'guid_url': guid_url,
         'tags': [tag._id for tag in file_.tags],
         'name': file_.name,
         'category': 'file',
